requirements.txt
passlib
pydantic-settings
python-jose
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
